report-repList-repView- refact
