+++ src/components/Reports/Reports.jsx
@@ -7,7 +7,10 @@ import nextArrowIcon from '../../images/next-arrow.svg';
 import { ArrowIcon, ButtonSwitch, Switch, SwitchData } from '../MonthPicker/MonthPicker.styled';
 import { ReportsContainer } from './Reports.styled'
 
-export const Reports = ({ switchData, clickOnSwitch }) => {
+export const Reports = ({ allCategories, switchData, clickOnSwitch }) => {
+
+    const costsType = allCategories.filter(item => item.type === "costs");
+    const incomeType = allCategories.filter(item=> item.type === "income")
 
     return (
         <ReportsContainer>
@@ -26,7 +29,11 @@ export const Reports = ({ switchData, clickOnSwitch }) => {
                 </ButtonSwitch>
             </Switch>
 
-            <ReportList category={switchData} />
+            {switchData === 'costs' ?
+                (<ReportList category={switchData} transactions={ costsType }/>)
+                :
+                (<ReportList category={switchData} transactions={ incomeType }/>)
+            }
             
         </ReportsContainer>
     )
