+++ src/views/ReportView/ReportView.jsx
@@ -7,7 +7,7 @@ import { ReportButton, MonthPicker, Reports, ReportStatistic } from '../../compo
 
 import { Chart } from '../../components/Chart'
 
-import {getAllTransactionsDATA} from '../../api/reportApi';
+// import {getAllTransactionsDATA} from '../../api/reportApi';
 
 import 'moment/locale/ru';
 import moment from 'moment';
@@ -18,6 +18,65 @@ import {
     ReportGraph
 } from './ReportView.styled'
 
+const transactions = [
+    {
+        "_id": "61e878e438630fdee30eaf62",
+        "type": "costs",
+        "category": "спорт,хобби",
+        "description": "gym",
+        "sum": 350,
+        "day": "09",
+        "month": "12",
+        "year": "2021",
+        "owner": "61e8754f36895bfe4aef28fc",
+        "createdAt": "2022-01-19T20:47:32.639Z",
+        "updatedAt": "2022-01-19T20:47:32.639Z",
+        "date": "09.12.2021"
+    },
+    {
+        "_id": "61e8791438630fdee30eaf67",
+        "type": "costs",
+        "category": "прочее",
+        "description": "killer",
+        "sum": 10000,
+        "day": "09",
+        "month": "12",
+        "year": "2021",
+        "owner": "61e8754f36895bfe4aef28fc",
+        "createdAt": "2022-01-19T20:48:20.311Z",
+        "updatedAt": "2022-01-19T20:48:20.311Z",
+        "date": "09.12.2021"
+    },
+    {
+        "_id": "61e8794e38630fdee30eaf6c",
+        "type": "costs",
+        "category": "коммуналка, связь",
+        "description": "water",
+        "sum": 500,
+        "day": "09",
+        "month": "12",
+        "year": "2021",
+        "owner": "61e8754f36895bfe4aef28fc",
+        "createdAt": "2022-01-19T20:49:18.125Z",
+        "updatedAt": "2022-01-19T20:49:18.125Z",
+        "date": "09.12.2021"
+    },
+    {
+        "_id": "61e8795638630fdee30eaf71",
+        "type": "costs",
+        "category": "коммуналка, связь",
+        "description": "water",
+        "sum": 512,
+        "day": "09",
+        "month": "12",
+        "year": "2021",
+        "owner": "61e8754f36895bfe4aef28fc",
+        "createdAt": "2022-01-19T20:49:26.964Z",
+        "updatedAt": "2022-01-19T20:49:26.964Z",
+        "date": "09.12.2021"
+    }
+];
+
 const ReportView = () => {
     const isMobile = useMediaQuery({ minWidth: 320, maxWidth: 767 });
     const isTabletOrDesktop = useMediaQuery({ minWidth: 768 });
@@ -26,16 +85,17 @@ const ReportView = () => {
     const [dateMonth, setDateMonth] = useState(moment(new Date()).format('MM'));
     const [dateYears, setDateYears] = useState(moment(new Date()).format('YYYY'));
     const [switchData, setSwitchData] = useState('costs');
-    const [categoriesCosts, setCategoriesCosts] = useState([]);
+    const [allCategories, setAllCategories] = useState([]);
 
 
     useEffect(() => {
-        async function getAllTransactions() {
-            const transactionsDATA = await getAllTransactionsDATA( dateYears, dateMonth, switchData);
-            setCategoriesCosts(transactionsDATA);
-            console.log(transactionsDATA.data);
-        }
-        getAllTransactions();
+        setAllCategories(transactions);
+        // async function getAllTransactions() {
+        //     const transactionsDATA = await getAllTransactionsDATA( dateYears, dateMonth, switchData);
+        //     setAllCategories(transactionsDATA);
+        //     console.log(transactionsDATA.data);
+        // }
+        // getAllTransactions();
     }, [switchData, dateMonth, dateYears]);
 
     const switchMonthLeft = () => {
@@ -98,6 +158,7 @@ const ReportView = () => {
                 </ReportStatistic>
 
                 <Reports
+                    allCategories={allCategories}
                     switchData={switchData}
                     clickOnSwitch={clickOnSwitch}
                 />
